Update Filament packages to v3.1.28
This commit updates the Filament libraries used by the project from version v3.1.27 to v3.1.28. As a result, the corresponding reference points in composer.lock file have been changed to reflect the new version's source. In addition, the timestamp of the package updates is also updated to reflect the current changes.